Make System.Net.WebSockets target ns1.7
