+++ src/System.Net.WebSockets/ref/System.Net.WebSockets.csproj
@@ -3,11 +3,10 @@
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <NuGetTargetMoniker>.NETStandard,Version=v1.3</NuGetTargetMoniker>
+    <NuGetTargetMoniker>.NETStandard,Version=v1.7</NuGetTargetMoniker>
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="System.Net.WebSockets.cs" />
-    <Compile Include="System.Net.WebSockets.Manual.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
