+++ src/System.Net.WebSockets/ref/project.json
@@ -1,18 +1,14 @@
 {
   "dependencies": {
-    "Microsoft.Win32.Primitives": "4.0.0",
-    "System.Collections.Specialized": "4.0.0",
-    "System.IO":  "4.0.0",
-    "System.Net.Primitives": "4.0.0",
-    "System.Runtime": "4.0.0",
-    "System.Security.Principal": "4.0.0",
-    "System.Threading.Tasks": "4.0.0"
+    "Microsoft.Win32.Primitives": "4.4.0-beta-24619-02",
+    "System.Collections.Specialized": "4.4.0-beta-24619-02",
+    "System.IO": "4.4.0-beta-24619-02",
+    "System.Net.Primitives": "4.4.0-beta-24619-02",
+    "System.Runtime": "4.4.0-beta-24619-02",
+    "System.Security.Principal": "4.4.0-beta-24619-02",
+    "System.Threading.Tasks": "4.4.0-beta-24619-02"
   },
   "frameworks": {
-    "netstandard1.3": {
-      "imports": [
-        "dotnet5.4"
-      ]
-    }
+    "netstandard1.7": {}
   }
 }
