+++ src/System.Net.WebSockets/src/System.Net.WebSockets.builds
@@ -4,7 +4,7 @@
   <ItemGroup>
     <Project Include="System.Net.WebSockets.csproj" />
     <Project Include="System.Net.WebSockets.csproj">
-      <TargetGroup>net46</TargetGroup>
+      <TargetGroup>net463</TargetGroup>
     </Project>
   </ItemGroup>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.traversal.targets))\dir.traversal.targets" />
