+++ src/System.Net.WebSockets/src/project.json
@@ -1,23 +1,21 @@
 {
   "frameworks": {
-    "netstandard1.3": {
+    "netstandard1.7": {
       "dependencies": {
-        "Microsoft.NETCore.Platforms": "1.0.1",
-        "Microsoft.Win32.Primitives": "4.0.0",
-        "System.Collections.Specialized": "4.0.0",
-        "System.Diagnostics.Debug": "4.0.0",
-        "System.Diagnostics.Tools": "4.0.0",
-        "System.Net.Primitives": "4.0.0",
-        "System.Resources.ResourceManager": "4.0.0",
-        "System.Runtime": "4.0.20",
-        "System.Security.Principal": "4.0.0",
-        "System.Threading.Tasks": "4.0.10"
-      },
-      "imports": [
-        "dotnet5.4"
-      ]
+        "Microsoft.Win32.Primitives": "4.4.0-beta-24619-02",
+        "System.Collections.Specialized": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Debug": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Tools": "4.4.0-beta-24619-02",
+        "System.Net.Primitives": "4.4.0-beta-24619-02",
+        "System.Resources.ResourceManager": "4.4.0-beta-24619-02",
+        "System.Runtime": "4.4.0-beta-24619-02",
+        "System.Runtime.Extensions": "4.4.0-beta-24619-02",
+        "System.Runtime.InteropServices": "4.4.0-beta-24619-02",
+        "System.Security.Principal": "4.4.0-beta-24619-02",
+        "System.Threading.Tasks": "4.4.0-beta-24619-02"
+      }
     },
-    "net46": {
+    "net463": {
       "dependencies": {
         "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
       }
