+++ src/System.Net.WebSockets/tests/System.Net.WebSockets.Tests.builds
@@ -5,6 +5,11 @@
     <Project Include="System.Net.WebSockets.Tests.csproj" />
     <Project Include="System.Net.WebSockets.Tests.csproj">
       <OSGroup>Windows_NT</OSGroup>
+      <TestTFMs>net463</TestTFMs>
+    </Project>
+    <Project Include="System.Net.WebSockets.Tests.csproj">
+      <TargetGroup>netstandard1.3</TargetGroup>
+      <OSGroup>Windows_NT</OSGroup>
       <TestTFMs>netcore50;net46</TestTFMs>
     </Project>
   </ItemGroup>
